Add flag to disable batching
